+++ src/common/utils/apiUtils.ts
@@ -2,7 +2,7 @@ import { uxLog } from ".";
 import * as c from "chalk";
 import { Connection, SfdxError } from "@salesforce/core";
 import { RestApiOptions, RecordResult } from "jsforce";
-import ora = require("ora");
+import * as ora from "ora";
 
 // Perform simple SOQL query (max results: 10000)
 export function soqlQuery(soqlQuery: string, conn: Connection): Promise<any> {